keep state in main module
