+++ tortoise/verifying.go
@@ -180,86 +180,3 @@ func (v *verifying) verify(logger log.Log, lid types.LayerID) bool {
 	}
 	return false
 }
-
-func decodeExceptions(logger log.Log, config Config, state *state, base, ballot *ballotInfo, exceptions *types.Votes) {
-	ballot.goodness |= base.goodness
-	if _, exist := state.badBeaconBallots[ballot.id]; exist {
-		ballot.goodness |= conditionBadBeacon
-	}
-	// inherit opinion from the base ballot by copying votes in range
-	votes := base.votes.Copy()
-	// add opinions from the local state [base layer, ballot layer)
-	for lid := base.layer; lid.Before(ballot.layer); lid = lid.Add(1) {
-		layer := state.layer(lid)
-		lvote := layerVote{
-			layerInfo: layer,
-			vote:      against,
-		}
-		for _, block := range layer.blocks {
-			lvote.blocks = append(lvote.blocks, blockVote{
-				blockInfo: block,
-				vote:      against,
-			})
-		}
-		votes.Append(&lvote)
-	}
-	// update exceptions
-	ballot.votes = votes
-	for _, bid := range exceptions.Support {
-		block, exist := state.blockRefs[bid]
-		if !exist {
-			ballot.goodness |= conditionVotesBeforeBase
-			continue
-		}
-		if block.layer.Before(base.layer) {
-			ballot.goodness |= conditionVotesBeforeBase
-		}
-		ballot.updateBlockVote(block, support)
-	}
-	for _, bid := range exceptions.Against {
-		block, exist := state.blockRefs[bid]
-		if !exist {
-			ballot.goodness |= conditionVotesBeforeBase
-			continue
-		}
-		if block.layer.Before(base.layer) {
-			ballot.goodness |= conditionVotesBeforeBase
-		}
-		ballot.updateBlockVote(block, against)
-	}
-	for _, lid := range exceptions.Abstain {
-		ballot.goodness |= conditionAbstained
-		if lid.Before(base.layer) {
-			ballot.goodness |= conditionVotesBeforeBase
-		}
-		ballot.updateLayerVote(lid, abstain)
-		layer := state.layer(lid)
-		layer.verifying.abstained = layer.verifying.abstained.Add(ballot.weight)
-	}
-	validateConsistency(config, state, ballot)
-	logger.With().Debug("decoded votes for ballot",
-		ballot.id,
-		ballot.layer,
-		log.Stringer("base", ballot.base.id),
-		log.Uint32("base layer", ballot.base.layer.Value),
-		log.Bool("base good", base.goodness.isGood()),
-		log.Bool("ignored", ballot.goodness.ignored()),
-		log.Bool("consistent", !ballot.goodness.notConsistent()),
-	)
-}
-
-func validateConsistency(config Config, state *state, ballot *ballotInfo) bool {
-	for lvote := ballot.votes.Tail; lvote != nil; lvote = lvote.prev {
-		if lvote.lid.Before(ballot.base.layer) {
-			break
-		}
-		for j := range lvote.blocks {
-			local, _ := getLocalVote(state, config, lvote.blocks[j].blockInfo)
-			if vote := lvote.blocks[j].vote; vote != local {
-				ballot.goodness |= conditionNotConsistent
-				return false
-			}
-		}
-	}
-	return true
-}
